+++ test/bigtest/interactors/settings/hrid-handling/hrid-handling.js
@@ -7,6 +7,8 @@ import {
 import ButtonInteractor from '@folio/stripes-components/lib/Button/tests/interactor';
 // eslint-disable-next-line
 import TextFieldInteractor from '@folio/stripes-components/lib/TextField/tests/interactor';
+import CalloutInteractor from '@folio/stripes-components/lib/Callout/tests/interactor';
+import ConfirmationModalInteractor from '@folio/stripes-components/lib/ConfirmationModal/tests/interactor';
 
 @interactor
 class StartWithFieldInteractor {
@@ -21,12 +23,14 @@ class AssignPrefixFieldInteractor {
 }
 
 @interactor class HRIDHandlingInteractor {
+  static defaultScope = '[data-test-hrid-handling-form]';
+
   submitFormButton = new ButtonInteractor('[data-test-submit-button]');
   submitFormButtonDisabled = property('[data-test-submit-button]', 'disabled');
   startWithFields = new StartWithFieldInteractor();
   assignPrefixFields = new AssignPrefixFieldInteractor();
-
-  static defaultScope = '[data-test-hrid-handling-form]';
+  callout = new CalloutInteractor();
+  confirmationModal = new ConfirmationModalInteractor('#confirm-edit-hrid-settings-modal');
 }
 
 export default new HRIDHandlingInteractor();