+++ test/bigtest/mocks/hridSettings.js
@@ -1,0 +1,17 @@
+const hridSettings = {
+  id: '65f6d006-393f-423b-9d2a-977d33371ab8',
+  instances: {
+    prefix: 'in',
+    startNumber: 1,
+  },
+  holdings: {
+    prefix: 'ho',
+    startNumber: 1,
+  },
+  items: {
+    prefix: 'it',
+    startNumber: 1,
+  },
+};
+
+export default hridSettings;