+++ test/bigtest/network/scenarios/update-hrid-settings-error.js
@@ -1,0 +1,3 @@
+export default server => {
+  server.put('/hrid-settings-storage/hrid-settings', { errors: [{ title: 'Internal server error, e.g. due to misconfiguration' }] }, 500);
+};