+++ test/bigtest/network/scenarios/update-hrid-settings-success.js
@@ -1,0 +1,3 @@
+export default server => {
+  server.put('/hrid-settings-storage/hrid-settings', {}, 204);
+};